# Do not store ar object for inventory refresh