--- a/app/models/manager_refresh/inventory_object.rb
+++ b/app/models/manager_refresh/inventory_object.rb
@@ -87,6 +87,10 @@ module ManagerRefresh
       inventory_collection.base_class_name
     end
 
+    def dependency?
+      !inventory_collection.saved?
+    end
+
     private
 
     def association?(inventory_collection_scope, key)