--- a/app/models/manager_refresh/inventory_object_lazy.rb
+++ b/app/models/manager_refresh/inventory_object_lazy.rb
@@ -28,7 +28,7 @@ module ManagerRefresh
     def dependency?
       # If key is not set, InventoryObjectLazy is a dependency, cause it points to the record itself. Otherwise
       # InventoryObjectLazy is a dependency only if it points to an attribute which is a dependency or a relation.
-      !!(!key || transitive_dependency?)
+      !!(!key || transitive_dependency?) && !inventory_collection.saved?
     end
 
     def transitive_dependency?