--- a/app/models/manager_refresh/save_collection/helper.rb
+++ b/app/models/manager_refresh/save_collection/helper.rb
@@ -55,8 +55,6 @@ module ManagerRefresh::SaveCollection
         record_index[inventory_collection.object_index_with_keys(unique_index_keys, record)] = record
       end
 
-      entity_builder = get_entity_builder(inventory_collection, association)
-
       inventory_collection_size = inventory_collection.size
       created_counter           = 0
       _log.info("*************** PROCESSING #{inventory_collection} of size #{inventory_collection_size} ***************")
@@ -66,7 +64,7 @@ module ManagerRefresh::SaveCollection
           record = record_index.delete(inventory_object.manager_uuid)
           if record.nil?
             next unless inventory_collection.create_allowed?
-            record          = entity_builder.create!(hash.except(:id))
+            record          = inventory_collection.model_class.create!(hash.except(:id))
             created_counter += 1
           else
             record.assign_attributes(hash.except(:id, :type))
@@ -94,15 +92,6 @@ module ManagerRefresh::SaveCollection
           deletes.map(&inventory_collection.delete_method)
         end
         _log.info("*************** DELETED #{inventory_collection} ***************")
-      end
-    end
-
-    def get_entity_builder(inventory_collection, association)
-      if inventory_collection.parent && !inventory_collection.arel
-        association_meta_info = inventory_collection.parent.class.reflect_on_association(inventory_collection.association)
-        association_meta_info.options[:through].blank? ? association : inventory_collection.model_class
-      else
-        inventory_collection.model_class
       end
     end
   end