… spec/models/manager_refresh/save_inventory/spec_parsed_data.rb