Do not store AR objects when doing refresh. And do not create objects using association, but directly through class. This brings down memory requirements heavily.

Implements:
https://www.pivotaltracker.com/story/show/137741549

Benchmark with refreshing AWS env with 77k AWS public images

New refresh with this patch applied
-------------------------------------------

**1st refresh (doing create of images):**

Memory consumed: ~1250MB
Time processing saving: ~500s

**2nd and other refreshes (doing mostly update of images):**

Memory consumed: ~1476MB
Time processing the saving: ~200s

New refresh without this patch applied
------------------------------------------------

**1st refresh (doing create of images):**

Memory consumed: ~3597MB
Time processing saving: ~1100s

**2nd and other refreshes (doing mostly update of images):**

Memory consumed: 2306~MB
Time processing the saving: ~444s

Old refresh
--------------

(there is O(n^2) bug in rails 5.0.1 I have locally fixed)

**1st refresh (doing create of images):**

Memory consumed: ~3344MB
Time processing saving: ~1238s

**2nd and other refreshes (doing mostly update of images):**

Memory consumed: ~3817MB
Time processing the saving: ~1537s

Result
----------

We see huge improvement both in memory and time. The big memory hog is also in the parser, doing api request to AWS takes about 600MB in memory. I've already filled an issue to aws-sdk, so we might see another decrease here after.