@durandom 2017-01-19
@Ladas thats some awesome performance and memory improvements