@kbrock 2017-01-19
including the reformatted results

|1st refresh|memory consumed|time processing|
|---|---|---
|old|3344MB|1238s
|before|3597MB|1100s
|after| 1250MB| 500s
|change|52%|55%

|2nd refresh|memory consumed|time processing|
|---|---|---
|old|3817MB|1537s
|before|2306MB| 444s
|after|1476MB| 200s
|change|36%|55%